source-sqlserver: move populate-source-timestamps test out of shared package
Per review feedback: the test in go/capture/sqlserver/tests/capture.go is
shared between source-sqlserver and source-sqlserver-ct, but
populate_source_ts_ms is a source-sqlserver-only config field. Move
TestPopulateSourceTimestamps and the extractTsMsValues helper into the
source-sqlserver package where they belong.
